+++ src/test/groovy/org/mockftpserver/fake/filesystem/UnixDirectoryListingFormatterTest.groovy
@@ -21,6 +21,8 @@ import org.junit.jupiter.api.Test
 import org.mockftpserver.test.AbstractGroovyTestCase
 
 import java.text.SimpleDateFormat
+import java.time.Duration
+import java.time.Instant
 
 /**
  * Tests for UnixDirectoryListingFormatter
@@ -29,82 +31,94 @@ import java.text.SimpleDateFormat
  */
 class UnixDirectoryListingFormatterTest extends AbstractGroovyTestCase {
 
-    static final FILE_NAME = "def.txt"
-    static final FILE_PATH = "/dir/$FILE_NAME"
-    static final DIR_NAME = "etc"
-    static final DIR_PATH = "/dir/$DIR_NAME"
-    static final OWNER = 'owner123'
-    static final GROUP = 'group456'
-    static final SIZE = 11L
-    static final LAST_MODIFIED = new Date()
-    static final FILE_PERMISSIONS = new Permissions('rw-r--r--')
-    static final DIR_PERMISSIONS = new Permissions('rwxr-xr-x')
+    private static final String FILE_NAME = "def.txt"
+    private static final String FILE_PATH = "/dir/$FILE_NAME"
+    private static final String DIR_NAME = "etc"
+    private static final String DIR_PATH = "/dir/$DIR_NAME"
+    private static final String OWNER = 'owner123'
+    private static final String GROUP = 'group456'
+    private static final long SIZE = 11L
+    private static final Date LAST_MODIFIED_RECENT = new Date()
+    private static final Date LAST_MODIFIED_OLDER = Date.from(Instant.now().minus(Duration.ofDays(181)))
+    private static final Permissions FILE_PERMISSIONS = new Permissions('rw-r--r--')
+    private static final Permissions DIR_PERMISSIONS = new Permissions('rwxr-xr-x')
 
-    private formatter
-    private lastModifiedFormatted
-    private defaultLocale
+    private UnixDirectoryListingFormatter formatter = new UnixDirectoryListingFormatter()
+    private Locale defaultLocale = Locale.default
+    private String lastModifiedRecentFormatted
+    private String lastModifiedOlderFormatted
 
     // "-rw-rw-r--    1 ftp      ftp           254 Feb 23  2007 robots.txt"
     // "-rw-r--r--    1 ftp      ftp      30014925 Apr 15 00:19 md5.sums.gz"
-    // "-rwxr-xr-x   1 c096336  iawebgrp    5778 Dec  1  2005 FU_WyCONN_updateplanaccess.sql"
+    // "-rwxr-xr-x   1 henry    users       5778 Dec  1  2005 planaccess.sql"
     // "drwxr-xr-x   2 c096336  iawebgrp    8192 Nov  7  2006 tmp"
     // "drwxr-xr-x   39 ftp      ftp          4096 Mar 19  2004 a"
 
     @Test
-    void testFormat_File() {
-        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED,
+    void testFormat_File_Recent() {
+        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED_RECENT,
                 owner: OWNER, group: GROUP, permissions: FILE_PERMISSIONS)
-        LOG.info(fileSystemEntry.toString())
-        verifyFormat(fileSystemEntry, "-rw-r--r--  1 owner123 group456              11 $lastModifiedFormatted def.txt")
+        verifyFormat(fileSystemEntry, "-rw-r--r--  1 owner123 group456              11 $lastModifiedRecentFormatted def.txt")
+    }
+
+    @Test
+    void testFormat_File_Older() {
+        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED_OLDER,
+                owner: OWNER, group: GROUP, permissions: FILE_PERMISSIONS)
+        verifyFormat(fileSystemEntry, "-rw-r--r--  1 owner123 group456              11 $lastModifiedOlderFormatted def.txt")
     }
 
     @Test
     void testFormat_File_Defaults() {
-        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED)
-        LOG.info(fileSystemEntry.toString())
-        verifyFormat(fileSystemEntry, "-rwxrwxrwx  1 none     none                  11 $lastModifiedFormatted def.txt")
+        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED_RECENT)
+        verifyFormat(fileSystemEntry, "-rwxrwxrwx  1 none     none                  11 $lastModifiedRecentFormatted def.txt")
     }
 
     @Test
     void testFormat_File_NonEnglishDefaultLocale() {
         Locale.setDefault(Locale.GERMAN)
-        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED)
-        LOG.info(fileSystemEntry.toString())
-        verifyFormat(fileSystemEntry, "-rwxrwxrwx  1 none     none                  11 $lastModifiedFormatted def.txt")
+        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED_RECENT)
+        verifyFormat(fileSystemEntry, "-rwxrwxrwx  1 none     none                  11 $lastModifiedRecentFormatted def.txt")
     }
 
     @Test
     void testFormat_File_NonEnglishLocale() {
         formatter.setLocale(Locale.FRENCH)
-        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED)
+        def fileSystemEntry = new FileEntry(path: FILE_PATH, contents: '12345678901', lastModified: LAST_MODIFIED_OLDER)
         LOG.info(fileSystemEntry.toString())
-        def dateFormat = new SimpleDateFormat(UnixDirectoryListingFormatter.DATE_FORMAT, Locale.FRENCH)
-        def formattedDate = dateFormat.format(LAST_MODIFIED)
+        def dateFormat = new SimpleDateFormat(UnixDirectoryListingFormatter.DATE_FORMAT_YEAR, Locale.FRENCH)
+        def formattedDate = dateFormat.format(LAST_MODIFIED_OLDER)
         def result = formatter.format(fileSystemEntry)
         assert result.contains(formattedDate)
     }
 
     @Test
-    void testFormat_Directory() {
-        def fileSystemEntry = new DirectoryEntry(path: DIR_PATH, lastModified: LAST_MODIFIED,
+    void testFormat_Directory_Recent() {
+        def fileSystemEntry = new DirectoryEntry(path: DIR_PATH, lastModified: LAST_MODIFIED_RECENT,
                 owner: OWNER, group: GROUP, permissions: DIR_PERMISSIONS)
-        LOG.info(fileSystemEntry.toString())
-        verifyFormat(fileSystemEntry, "drwxr-xr-x  1 owner123 group456               0 $lastModifiedFormatted etc")
+        verifyFormat(fileSystemEntry, "drwxr-xr-x  1 owner123 group456               0 $lastModifiedRecentFormatted etc")
+    }
+
+    @Test
+    void testFormat_Directory_Older() {
+        def fileSystemEntry = new DirectoryEntry(path: DIR_PATH, lastModified: LAST_MODIFIED_OLDER,
+                owner: OWNER, group: GROUP, permissions: DIR_PERMISSIONS)
+        verifyFormat(fileSystemEntry, "drwxr-xr-x  1 owner123 group456               0 $lastModifiedOlderFormatted etc")
     }
 
     @Test
     void testFormat_Directory_Defaults() {
-        def fileSystemEntry = new DirectoryEntry(path: DIR_PATH, lastModified: LAST_MODIFIED)
-        LOG.info(fileSystemEntry.toString())
-        verifyFormat(fileSystemEntry, "drwxrwxrwx  1 none     none                   0 $lastModifiedFormatted etc")
+        def fileSystemEntry = new DirectoryEntry(path: DIR_PATH, lastModified: LAST_MODIFIED_RECENT)
+        verifyFormat(fileSystemEntry, "drwxrwxrwx  1 none     none                   0 $lastModifiedRecentFormatted etc")
     }
 
     @BeforeEach
     void setUp() {
-        formatter = new UnixDirectoryListingFormatter()
-        def dateFormat = new SimpleDateFormat(UnixDirectoryListingFormatter.DATE_FORMAT, Locale.ENGLISH)
-        lastModifiedFormatted = dateFormat.format(LAST_MODIFIED)
-        defaultLocale = Locale.default
+        def dateFormatRecent = new SimpleDateFormat(UnixDirectoryListingFormatter.DATE_FORMAT_HOURS_MINUTES, Locale.ENGLISH)
+        lastModifiedRecentFormatted = dateFormatRecent.format(LAST_MODIFIED_RECENT)
+
+        def dateFormatOlder = new SimpleDateFormat(UnixDirectoryListingFormatter.DATE_FORMAT_YEAR, Locale.ENGLISH)
+        lastModifiedOlderFormatted = dateFormatOlder.format(LAST_MODIFIED_OLDER)
     }
 
     @AfterEach
@@ -113,6 +127,7 @@ class UnixDirectoryListingFormatterTest extends AbstractGroovyTestCase {
     }
 
     private void verifyFormat(FileSystemEntry fileSystemEntry, String expectedResult) {
+        LOG.info(fileSystemEntry.toString())
         def result = formatter.format(fileSystemEntry)
         LOG.info("result=  [$result]")
         LOG.info("expected=[$expectedResult]")